back_py/requirements.txt
Django==5.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg[binary]>=3.1,<3.3
kubernetes==31.0.0
python-dotenv==1.0.1
gunicorn==23.0.0
celery==5.4.0
redis==5.1.1
docker==7.1.0
